feat(CodeEditor): implement useFoldGutterExtension for code folding functionality
